Move ComposableFirmwareUpdaterError to separate file
